refactor(test): extract elicitation tests into ElicitationIntegrationTestCase
Move form-mode and URL-mode elicitation integration tests out of
MCPServerIntegrationTestCase into a dedicated ElicitationIntegrationTestCase.
Extract shared SSE infrastructure into AbstractMCPIntegrationTestCase base class.
